refactor: move 11 pure capability/config types to homeboy-extension-contract
Sixth sub-cut of the extension-contract campaign (Stage 1, wiki 12840).

Extracts a closed cluster of 11 pure serde types from manifest.rs into a new
contract manifest_capability_config module: ScriptsConfig, ReleasePreflightConfig,
RuntimeRequirementsConfig, RuntimeRequirementConfig, AgentRuntimeManifestConfig,
ProvidesConfig, DiscoveryMarkerConfig, DocTarget, ComponentEnvConfig,
FeatureContextRule, ExtensionToolDiagnosticDeclaration.

All 11 have zero impls, no custom serde helpers, and reference only primitives
or each other within the set. manifest.rs re-exports them (pub use), so the
existing pub use manifest::{...} surface in mod.rs is unchanged and every
crate::extension::* consumer keeps working. manifest.rs drops from 27 to 16
top-level types.
