+++ crates/homeboy-extension-contract/src/lib.rs
@@ -13,6 +13,7 @@ pub mod action_types;
 pub mod core_compat;
 pub mod exec_context;
 pub mod manifest_action_config;
+pub mod manifest_capability_config;
 pub mod manifest_deploy_config;
 pub mod manifest_test_config;
 pub mod runner_contract;
